requirements.txt
streamlit
crewai
crewai-tools
langchain-google-genai
fpdf
python-dotenv
